[ refactor ] UserQueryRepository 재설정  (#44 )
