+++ auth/src/main/java/com/linkeleven/msa/auth/domain/repository/UserRepository.java
@@ -2,6 +2,10 @@ package com.linkeleven.msa.auth.domain.repository;
 
 import java.util.Optional;
 
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Slice;
+
+import com.linkeleven.msa.auth.application.dto.UserQueryResponseDto;
 import com.linkeleven.msa.auth.domain.model.User;
 
 public interface UserRepository {
@@ -12,4 +16,6 @@ public interface UserRepository {
 	boolean existsByUsername(String username);
 
 	User save(User user);
+
+	Slice<UserQueryResponseDto> findUserByUsername(String username, Pageable pageable);
 }
