+++ auth/src/main/java/com/linkeleven/msa/auth/infrastructure/repository/UserQueryRepository.java
@@ -1,10 +1,10 @@
-package com.linkeleven.msa.auth.domain.repository;
+package com.linkeleven.msa.auth.infrastructure.repository;
 
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 
 import com.linkeleven.msa.auth.application.dto.UserQueryResponseDto;
 
-public interface UserRepositoryCustom {
+public interface UserQueryRepository {
 	Slice<UserQueryResponseDto> findUserByUsername(String username, Pageable pageable);
 }
